Linux changes

+++ scripts/burn.sh
@@ -5,12 +5,17 @@ if [ -z "$1" ]; then
 	exit 1
 fi
 
+if ! [ -f "$1" ]; then
+	echo "File not found: $1"
+	exit 1
+fi
+
 disks=""
 
 if [ -x "$(command -v diskutil)" ]; then
 	disks=$(diskutil list | awk '/external/ {print $1}')
 elif [ -x "$(command -v lsblk)" ]; then
-	disks=$(lsblk -o NAME,TRAN | awk '/usb/{print $1}')
+	disks=$(lsblk -o NAME,TRAN | awk '/usb/{print "/dev/"$1}')
 else
 	echo "I don't know what tool to identify your USB drives with."
 	exit 1
@@ -22,5 +27,10 @@ if [ -z "$disks" ]; then
 fi
 
 disk=$(fzf <<<"$disks")
+
+if [ -z "$disk" ]; then
+	echo "No USB drive selected."
+	exit 1
+fi
 
 sudo dd if="$1" of="$disk" bs=1M status=progress
